+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased changes
 
 - Add `Moar.Assertions.assert_contains/2`
+- Soft-deprecate `Moar.Assertions.assert_eq`'s `ignore_whitespace: :leading_and_trailing` option in
+  favor of `whitespace: :squish` and `whitespace: :trim`
 
 ## 1.31.0